Decouple DashVariant from icons and UITools
